[HIP] search fatbin symbols for libs passed by -l (#104638)
For -fgpu-rdc linking, clang needs to collect undefined fatbin symbols
and resolve them to the embedded fatbin.

This has been done for object files and archive files passed as input
files to clang.

However, the same action is not performed for archive files passed
through -l options, which causes missing symbols.

This patch adds that.

Change-Id: I642cb33bbfc16eadb47261ee4ec8d9038d9f15cc