[Comgr] Add initial SPIR-V translation support
In this commit, we add a new SPIR-V data type:

  AMD_COMGR_DATA_KIND_SPIRV

and translation action:

  AMD_COMGR_ACTION_TRANSLATE_SPIRV_TO_BC

With this data type and action, users can provide a set of SPIR-V (.spv)
files, and have Comgr translate them into a set of LLVM IR bitcode (.bc)
objects.

We also add a HIP and OpenCL test for the new translation action. The
HIP test is currently disabled until compiler support is finalized.

Change-Id: Ic05c880158b9c4fd49550a5b2e51bb26dfed0efc