[Comgr] Add new DataAction flag for device library linking
Previously, some Comgr actions implicitly linked the device libraries,
while others did not.

With this new DataAction option, users can explicitly enforce or disable
device library linking via:

  amd_comgr_action_info_set_device_lib_linking(
                    DataAction ActionInfo,
                    bool ShouldLinkDeviceLibs)

This change impacts the following actions:

  AMD_COMGR_ACTION_COMPILE_SOURCE_TO_BC
  AMD_COMGR_ACTION_COMPILE_SOURCE_TO_EXECUTABLE
  AMD_COMGR_ACTION_COMPILE_SOURCE_TO_RELOCATABLE

  AMD_COMGR_ACTION_CODEGEN_BC_TO_RELOCATABLE
  AMD_COMGR_ACTION_CODEGEN_BC_TO_ASSEMBLY
  AMD_COMGR_ACTION_LINK_RELOCATABLE_TO_EXECUTABLE

If device library linking is required for one of the above actions, it
should be explicitly enforced via the new API. Currently the default
behavior of each action is preserved, but in the future the default
will be changed to NOT link, and linking will only occur if
amd_comgr_action_info_set_device_lib_linking() is used.

Note: In Comgr 4.0, we plan to remove

  AMD_COMGR_COMPILE_SOURCE_WITH_DEVICE_LIBS_TO_BC

In favor of

  amd_comgr_action_info_set_device_lib_linking(ActionInfo, true);
  AMD_COMGR_COMPILE_SOURCE_TO_BC

Change-Id: I46411d1bf01354ac5ad434813df84569956a34d5