[Clang][OpenMP] Do not use feature option during packaging and unpackaging
Clang-offload-packager allows packaging and unpackaging of images
based on an arbitrary list of key-value pairs where only
triple-key is mandatory.

Using target features as a key during packaging and unpackaging
is not correct, as clang does not allow packaging multiple
images in one binary which only differ in a target feature.
TargetID features (xnack and sramecc) anyways are handled using
arch-key and not as target features.

Change-Id: I17d9bdc5fce68101af10304cc0be391e7e392448