[Comgr] Add LIT testing capabilites to Comgr
To robustly test Comgr, we can leverage LLVM tools like llvm-dis,
llvm-link, lld, etc.

To facilitate this, in this patch we set up the infrastructure
for Comgr llvm-lit tests.

These tests can be added in the comgr/test-lit directory, and can
be run via:

  make test-lit

They are also added to the check-comgr target, so they can be run
along with the traditional tests with

  make check-comgr

Change-Id: I18fafd09c1ea280d78d32071e6877868cfaa66e5