Fix spilling/occupancy metrics, remove added subtarget functions
Excess RP (whether for spilling reduction or occupancy increase) is now
computed using pre-existing functions on the subtarget, instead of using
custom functions that end up more or less duplicating the behavior of
others. In addition, the calculation now takes into account the
occupancy bounds derived from the range of flat workgroup sizes as well
as a potential user-provided maximum number of VGPRs to use.

New unit tests are added to check for some of the edge cases arising
from custom attribute values.