[InferAddressSpaces] Add support for maintaining DIOp-based DIExpressions
In order to do this, add support for casting between pointer address
space types to replaceAllDbgUsesWith. We should probably upstream a
version of this, but for now this only affects DIOp-based DIExpressions
(to reduce the diff with upstream). InstCombine can also hit this case
when eliminating casts, see added test. Also, fix a bug in the verifier
where DIOp-DIExpressions appearing in debug records weren't checked.

This is still missing support in AsmPrinter for adding the correct
address space attribute on the variable's type DIE. AsmPrinter should
determine that type by finding a common address space on the result
types of the DIExpressions that make up that variable's DW_ATE_location.

Fixes SWDEV-527137.