[HeterogeneousDwarf] Support DWARF emission of DW_AT_LLVM_address_space
This is done by looking at the address space of each DIExpression that
makes up the loclist/location of a variable. If all the expressions
produce a pointer of the same address space, then use that address space
in the variable's type DIE.

This attribute isn't yet supported by rocgdb, but DW_AT_address_class
does have some support for pointer types.

This commit also introduces the notion of a "divergent" address spaces,
which is the case where a DIExpression's result type is a generic
pointer, but some optimization moved it to a specific address space. In
that case, AsmPrinter emits a pointer in the specific address space and
adds an attribute to the variable type DIE with the divergent address
space.

Fixes SWDEV-527137.